requirements.txt
pandas
datetime
numpy
faiss-cpu
openai
dotenv